+++ CL/main.py
@@ -3,12 +3,18 @@ import argparse
 import os
 import shutil
 import parsegen
+import log
+
 
 def print_help() -> None:
     print("py main.py [--parsegen] [--out=<outdir>] [--cc=<cc>]")
     print("\t--parsegen ... generates odin files out of cl headers (set by default)")
     print("\t--out      ... specifies output directory for *.odin bindings")
-    print("\t--cc       ... specifies C compiler to be used for CL header preprocessing")
+    print(
+        "\t--cc       ... specifies C compiler to be used for CL header preprocessing"
+    )
+    print("\t--verbose  ... enables DEBUG messages")
+
 
 def arg_validator(s: str) -> str:
     s = s.lower()
@@ -17,7 +23,10 @@ def arg_validator(s: str) -> str:
         case "clang" | "gcc" | "cl":
             return s
         case _:
-            raise argparse.ArgumentTypeError("Invalid value for --cc (expected: clang|gcc|cl).")
+            raise argparse.ArgumentTypeError(
+                "Invalid value for --cc (expected: clang|gcc|cl)."
+            )
+
 
 def find_cc(cc: str | None) -> str:
     if not cc:
@@ -38,10 +47,19 @@ def find_cc(cc: str | None) -> str:
 
 def main() -> None:
     parser = argparse.ArgumentParser(add_help=False)
-    parser.add_argument("--parsegen", action="store_true", help="Generates odin files out of cl headers")
-    parser.add_argument("--out", type=str, help="Specifies output directory of *.odin bindings")
-    parser.add_argument("--cc", type=arg_validator, help="Specifies C compiler to be used for CL header preprocessing")
+    parser.add_argument(
+        "--parsegen", action="store_true", help="Generates odin files out of cl headers"
+    )
+    parser.add_argument(
+        "--out", type=str, help="Specifies output directory of *.odin bindings"
+    )
+    parser.add_argument(
+        "--cc",
+        type=arg_validator,
+        help="Specifies C compiler to be used for CL header preprocessing",
+    )
     parser.add_argument("--help", action="store_true", help="Show help message")
+    parser.add_argument("--verbose", action="store_true", help="Enables DEBUG messages")
 
     args, unknown = parser.parse_known_args()
 
@@ -52,10 +70,14 @@ def main() -> None:
     elif args.help:
         print_help()
     else:
+        log.set_debug_msg_out(args.verbose)
+
         out_dir = args.out or os.getcwd()
         parsegen.main(find_cc(args.cc), out_dir)
 
+
 import sys
-sys.stdout.reconfigure(encoding='utf-8')
+
+sys.stdout.reconfigure(encoding="utf-8")
 if __name__ == "__main__":
     main()